requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
colorama==0.4.5
crispy-bulma==0.8.0
distlib==0.3.5
Django==4.0.6
django-ckeditor==6.5.0
django-crispy-forms==1.14.0
django-js-asset==2.0.0
filelock==3.7.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
virtualenv==20.15.1